extracting github ref parsing to a single place
