+++ apps/frontman_server/lib/frontman_server.ex
@@ -18,6 +18,7 @@ defmodule FrontmanServer do
     {Organizations, []},
     {Providers, []},
     PlayGithub,
+    PlayGithub.GithubReference,
     {Tasks, []},
     {Frameworks, []},
     BrandTokens,
