+++ apps/frontman_server/lib/frontman_server/play_github.ex
@@ -6,16 +6,11 @@
 
 defmodule FrontmanServer.PlayGithub do
   @moduledoc """
-  Parses GitHub-shaped paths routed through the local PlayGithub host.
+  Coordinates PlayGithub repository sandboxes.
   """
 
-  alias FrontmanServer.PlayGithub.{
-    Daytona,
-    GithubIssuePath,
-    GithubPath,
-    GithubRepositoryPath,
-    GithubTreePath
-  }
+  alias FrontmanServer.PlayGithub.{Daytona, GithubReference}
+  alias FrontmanServer.PlayGithub.GithubReference.{RepositoryPath, TreePath}
 
   @repo_url_label "frontman.playgithub.repo_url"
   @repo_cloned_label "frontman.playgithub.cloned"
@@ -49,100 +44,14 @@ defmodule FrontmanServer.PlayGithub do
 
   @type repository_command :: :create | :start | :clone | :install | :dev
 
-  @spec parse_path([String.t()]) ::
-          {:ok, GithubPath.t()}
-          | {:error, :missing_owner_or_repo}
-          | {:error, :missing_tree_ref}
-          | {:error, :invalid_issue_number}
-          | {:error, {:unsupported_github_path, [String.t()]}}
-  def parse_path([owner, repo]) do
-    {:ok,
-     %GithubPath{
-       owner: owner,
-       repo: repo,
-       resource: %GithubRepositoryPath{},
-       raw_segments: [owner, repo]
-     }}
-  end
-
-  def parse_path([_owner, _repo, "tree"]) do
-    {:error, :missing_tree_ref}
-  end
-
-  def parse_path([owner, repo, "tree", ref | path_segments]) do
-    {:ok,
-     %GithubPath{
-       owner: owner,
-       repo: repo,
-       resource: %GithubTreePath{ref: ref, path_segments: path_segments},
-       raw_segments: [owner, repo, "tree", ref | path_segments]
-     }}
-  end
-
-  def parse_path([_owner, _repo, "issues"]) do
-    {:error, :invalid_issue_number}
-  end
-
-  def parse_path([owner, repo, "issues", number]) do
-    case Integer.parse(number) do
-      {issue_number, ""} when issue_number > 0 ->
-        {:ok,
-         %GithubPath{
-           owner: owner,
-           repo: repo,
-           resource: %GithubIssuePath{number: issue_number},
-           raw_segments: [owner, repo, "issues", number]
-         }}
-
-      _ ->
-        {:error, :invalid_issue_number}
-    end
-  end
-
-  def parse_path([_owner, _repo, "issues", _number | _extra_segments]) do
-    {:error, :invalid_issue_number}
-  end
-
-  def parse_path([_owner]), do: {:error, :missing_owner_or_repo}
-  def parse_path([]), do: {:error, :missing_owner_or_repo}
-
-  def parse_path([_owner, _repo | extra_segments]) do
-    {:error, {:unsupported_github_path, extra_segments}}
-  end
-
-  @spec format_path(GithubPath.t()) :: String.t()
-  def format_path(
-        %GithubPath{
-          owner: owner,
-          repo: repo,
-          resource: resource,
-          raw_segments: raw_segments
-        } = github_path
-      ) do
-    [
-      "owner: #{owner}",
-      "repo: #{repo}",
-      "github_url: #{github_url(github_path)}",
-      format_resource(resource),
-      "raw_segments: #{Enum.join(raw_segments, "/")}"
-    ]
-    |> List.flatten()
-    |> Enum.join("\n")
-  end
-
-  @spec github_url(GithubPath.t()) :: String.t()
-  def github_url(%GithubPath{owner: owner, repo: repo}) do
-    "https://github.com/#{owner}/#{repo}"
-  end
-
-  @spec sandbox_name(GithubPath.t()) :: String.t()
-  def sandbox_name(%GithubPath{} = github_path) do
-    github_path
-    |> sandbox_identity()
+  @spec sandbox_name(GithubReference.t()) :: String.t()
+  def sandbox_name(%GithubReference{} = github_reference) do
+    github_reference
+    |> GithubReference.repository_identity()
     |> sandbox_name_for_identity()
   end
 
-  @spec run_repository_command(GithubPath.t(), repository_command(), keyword()) ::
+  @spec run_repository_command(GithubReference.t(), repository_command(), keyword()) ::
           {:ok,
            %{
              command: String.t(),
@@ -183,24 +92,24 @@ defmodule FrontmanServer.PlayGithub do
   def run_repository_command(github_path, command, opts \\ [])
 
   def run_repository_command(
-        %GithubPath{resource: %GithubRepositoryPath{}} = github_path,
+        %GithubReference{resource: %RepositoryPath{}} = github_reference,
         command,
         opts
       )
       when command in [:create, :start, :clone, :install, :dev] do
-    run_repository_target_command(repository_target(github_path), command, opts)
+    run_repository_target_command(repository_target(github_reference), command, opts)
   end
 
   def run_repository_command(
-        %GithubPath{resource: %GithubTreePath{}} = github_path,
+        %GithubReference{resource: %TreePath{}} = github_reference,
         command,
         opts
       )
       when command in [:create, :start, :clone, :install, :dev] do
-    run_repository_target_command(repository_target(github_path), command, opts)
+    run_repository_target_command(repository_target(github_reference), command, opts)
   end
 
-  def run_repository_command(%GithubPath{}, command, _opts)
+  def run_repository_command(%GithubReference{}, command, _opts)
       when command in [:create, :start, :clone, :install, :dev] do
     {:error, :not_repository_path}
   end
@@ -229,36 +138,24 @@ defmodule FrontmanServer.PlayGithub do
     })
   end
 
-  defp repository_target(%GithubPath{} = github_path) do
-    repo_url = github_url(github_path)
+  defp repository_target(%GithubReference{} = github_reference) do
+    repo_url = GithubReference.github_url(github_reference)
 
-    github_path
+    github_reference
     |> resource_target()
     |> Map.merge(%{
       repo_url: repo_url,
-      sandbox_name: sandbox_name_for_identity(sandbox_identity(github_path))
+      sandbox_name:
+        sandbox_name_for_identity(GithubReference.repository_identity(github_reference))
     })
   end
 
-  defp sandbox_identity(%GithubPath{resource: %GithubTreePath{} = tree_path} = github_path) do
-    repo_url = github_url(github_path)
-
-    case GithubTreePath.repo_path(tree_path) do
-      nil -> "#{repo_url}/tree/#{tree_path.ref}"
-      repo_path -> "#{repo_url}/tree/#{tree_path.ref}/#{repo_path}"
-    end
-  end
-
-  defp sandbox_identity(%GithubPath{} = github_path) do
-    github_url(github_path)
-  end
-
-  defp resource_target(%GithubPath{resource: %GithubRepositoryPath{}}) do
+  defp resource_target(%GithubReference{resource: %RepositoryPath{}}) do
     %{github_ref: nil, repo_path: nil, workspace_path: "workspace"}
   end
 
-  defp resource_target(%GithubPath{resource: %GithubTreePath{} = tree_path}) do
-    repo_path = GithubTreePath.repo_path(tree_path)
+  defp resource_target(%GithubReference{resource: %TreePath{} = tree_path}) do
+    repo_path = TreePath.repo_path(tree_path)
 
     %{
       github_ref: tree_path.ref,
@@ -1446,26 +1343,4 @@ defmodule FrontmanServer.PlayGithub do
 
     "playgithub-#{hash}"
   end
-
-  defp format_resource(%GithubRepositoryPath{}) do
-    "resource: repository"
-  end
-
-  defp format_resource(%GithubTreePath{} = tree_path) do
-    [
-      "resource: tree",
-      "ref: #{tree_path.ref}",
-      "path: #{format_repo_path(GithubTreePath.repo_path(tree_path))}"
-    ]
-  end
-
-  defp format_resource(%GithubIssuePath{number: number}) do
-    [
-      "resource: issue",
-      "issue_number: #{number}"
-    ]
-  end
-
-  defp format_repo_path(nil), do: ""
-  defp format_repo_path(repo_path), do: repo_path
 end
