+++ apps/frontman_server/lib/frontman_server/play_github/github_reference/tree_path.ex
@@ -4,7 +4,7 @@
 # Licensed under the AGPL-3.0 — see LICENSE for details.
 # Additional terms apply — see AI-SUPPLEMENTARY-TERMS.md
 
-defmodule FrontmanServer.PlayGithub.GithubTreePath do
+defmodule FrontmanServer.PlayGithub.GithubReference.TreePath do
   @moduledoc """
   GitHub tree path with a string-only ref split.
   """
